+++ app/configurator/components/chart-controls/color-palette.tsx
@@ -39,7 +39,7 @@ import {
 import useEvent from "@/utils/use-event";
 import { useUserPalettes } from "@/utils/use-user-palettes";
 
-import { ConfiguratorDrawer } from "../drawer";
+import { ConfiguratorDrawer } from "../drawers";
 
 import { ColorPaletteDrawerContent } from "./drawer-color-palette-content";
