+++ app/configurator/components/layout.tsx
@@ -3,7 +3,7 @@ import { makeStyles } from "@mui/styles";
 import clsx from "clsx";
 
 import { HEADER_HEIGHT_CSS_VAR } from "@/components/header-constants";
-import { DRAWER_WIDTH } from "@/configurator/components/drawer";
+import { DRAWER_WIDTH } from "@/configurator/components/drawers";
 import { useResizeObserver } from "@/utils/use-resize-observer";
 
 export const LAYOUT_HEADER_HEIGHT = 88;
